refactor: clean up story helpers
