+++ packages/storybook/src/Chip/Chip.stories.tsx
@@ -25,8 +25,8 @@ import {
 } from '@monorail/components/icons'
 import { useTheme } from '@monorail/utils'
 
-import { capitalizeFirstLetter } from '../helpers/helpers.js'
-import { isMeteorTheme, story } from '../helpers/storybook.js'
+import { capitalizeFirstLetter, isMeteorTheme } from '../helpers/helpers.js'
+import { story } from '../helpers/storybook.js'
 
 export default { title: 'Data Display/Chip', component: Chip }
