+++ packages/storybook/src/helpers/storybook.ts
@@ -14,15 +14,7 @@ import type {
 import type { AnnotatedStoryFn } from '@storybook/types'
 
 import { meteorLightTheme } from '@monorail/themes'
-import { RawColor as ClassicDarkRawColors } from '@monorail/themes/classic/theme/dark'
-import { RawColor as ClassicLightRawColors } from '@monorail/themes/classic/theme/light'
-import { RawColor as MeteorDarkRawColors } from '@monorail/themes/meteor/theme/dark'
-import { RawColor as MeteorLightRawColors } from '@monorail/themes/meteor/theme/light'
-import { RawColor as MuiRawColors } from '@monorail/themes/mui/theme'
-import { RawColor as PcteDarkRawColors } from '@monorail/themes/pcte/theme/dark'
-import { RawColor as PcteLightRawColors } from '@monorail/themes/pcte/theme/light'
 
-import { ThemeName } from '../theme/palette/palette.types'
 import { isNonEmptyString } from './typeGuards.js'
 
 type A11yParameter = {
@@ -238,32 +230,3 @@ export function story<T extends DefaultArgs>(
 
   return NewStory
 }
-
-/**
- * Gets the `RawColor` object of a theme. Used for displaying the corresponding global token and value of an alias token.
- * @param themeName theme.palette.name
- * @returns RawColors object
- */
-export const getRawColorObject = (themeName: ThemeName) => {
-  switch (themeName) {
-    case ThemeName.MeteorLight:
-      return MeteorLightRawColors
-    case ThemeName.MeteorDark:
-      return MeteorDarkRawColors
-    case ThemeName.ClassicLight:
-      return ClassicLightRawColors
-    case ThemeName.ClassicDark:
-      return ClassicDarkRawColors
-    case ThemeName.MUILight:
-      return MuiRawColors
-    case ThemeName.MUIDark:
-      return MuiRawColors
-    case ThemeName.PCTELight:
-      return PcteLightRawColors
-    case ThemeName.PCTEDark:
-      return PcteDarkRawColors
-  }
-}
-
-export const isMeteorTheme = (themeName: ThemeName) =>
-  themeName === ThemeName.MeteorLight || themeName === ThemeName.MeteorDark
